[recipes] Clone `chromium/src` with no `--depth`
This change corrects our cloning strategy for Chromium, as the current
code was still passing `--depth`, and that can have a detrimental
performance impact when compared to a fully-shared checkout.

Bug: https://github.com/brave/brave-browser/issues/56748
