+++ tools/recipes/recipe_modules/chromium_checkout/examples/full.expected/existing_branch.json
@@ -132,6 +132,7 @@
     "cmd": [
       "git",
       "fetch",
+      "--no-show-forced-updates",
       "origin",
       "main"
     ],
