+++ tools/recipes/recipe_modules/chromium_checkout/examples/full.expected/existing_tag.json
@@ -132,6 +132,7 @@
     "cmd": [
       "git",
       "fetch",
+      "--no-show-forced-updates",
       "--no-tags",
       "origin",
       "refs/tags/151.0.7917.1:refs/tags/151.0.7917.1"
