+++ tools/recipes/recipe_modules/chromium_checkout/examples/full.expected/fresh_release_branch.json
@@ -171,6 +171,7 @@
     "cmd": [
       "git",
       "fetch",
+      "--no-show-forced-updates",
       "origin",
       "refs/branch-heads/6834"
     ],
