+++ tools/recipes/recipes/tools/ast_grep/package_ast_grep.expected/mac.json
@@ -107,8 +107,6 @@
       "--no-checkout",
       "--local",
       "--shared",
-      "--depth",
-      "1",
       "/b/cache/chromium.googlesource.com-chromium-src",
       "[WORKSPACE]/b/src"
     ],
