+++ tools/recipes/recipes/tools/ast_grep/package_ast_grep.py
@@ -23,7 +23,7 @@ PROPERTIES = InputProperties
 
 def RunSteps(api: RecipeScriptApi, properties: InputProperties) -> None:
     chromium_src = api.chromium_checkout.ensure_checkout(
-        ref=properties.chromium_ref, depth=1)
+        ref=properties.chromium_ref)
 
     brave_root = api.brave_core_checkout.deploy([
         'third_party/ast-grep',
